Do not inform user when the alpha identifier is a null data.
Current code is not in accordance with the spec
3GPP TS 11.14, 6.4.5. Play Tone, as user is informed when the
alpha identifier is a null data.

With reference to the spec, if there is no alpha identifier,
show the text message to the user depending on the config setting.

If alpha identifier data is null, play only tone and do not
inform the user.

Move the code to handle play tone from ToneDialog to StkAppService
in cases where only tone needs to be played and no user inform needed.

Dismiss the tone dialog from StkAppService after play tone finish.

Send the intent to StkAppService from ToneDialog to finish
the play tone, when user presses back key.

Ref spec: 3GPP TS 11.14, 6.4.5. Play Tone.

CRs-Fixed: 631723, 517578
Change-Id: I0acf11e86714138e6d3193b17181c6710d06dfb2